tutor_system_reference/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
pydantic==2.9.2
pyyaml==6.0.2
numpy>=1.26,<3
torch>=2.2.0
scikit-learn==1.5.2
